feat: add latest build

+++ backend/src/controllers/user.js
@@ -90,9 +90,14 @@ const changePassword = async (req, res) => {
 const deleteUser = async (req, res) => {
   const { id } = req.params;
 
+  if (!id) {
+    return res.status(400).json({ message: "User ID is required" });
+  }
+
   try {
     await userService.deleteUser(id);
-    res.status(204).json({ message: "User deleted" });
+    
+    res.status(200).json({ message: "User deleted" });
   } catch (error) {
     errorHandler(error, res);
   }
